+++ queries/table_creation.sql
@@ -1,46 +1,47 @@
 -- Dropping tables in case they already exist
 
 -- Child tables
-drop table dept_emp;
-drop table salaries;
-drop table dept_manager;
+drop table if exists dept_emp;
+drop table if exists salaries;
+drop table if exists dept_manager;
 
 -- Parent tables
-drop table departments;
-drop table employees;
-drop table titles;
+drop table if exists departments;
+drop table if exists employees;
+drop table if exists titles;
 
 -- Creating all tables
 create table titles (
-	title_id varchar(5) primary key,
-	title varchar (30));
+	title_id varchar(5) primary key not null,
+	title varchar (30) not null);
 
 create table employees (
-	emp_no int primary key,
-	emp_title_id varchar(5),
-	birth_date varchar(10),
-	first_name varchar(30),
-	last_name varchar(30),
-	sex varchar (1),
-	hire_date varchar(10),
-	constraint fk_titles foreign key(emp_title_id) references titles(title_id));
+	emp_no int primary key not null,
+	emp_title_id varchar(5) not null,
+	birth_date varchar(10) not null,
+	first_name varchar(30) not null,
+	last_name varchar(30) not null,
+	sex varchar (1) not null,
+	hire_date varchar(10) not null,
+	foreign key(emp_title_id) references titles(title_id));
 
 create table salaries (
-	emp_no int,
-	salary int,
-	constraint fk_employee foreign key(emp_no) references employees(emp_no));
+	emp_no int not null,
+	salary int not null,
+	foreign key(emp_no) references employees(emp_no));
 
 create table departments (
-	dept_no varchar(4) primary key,
-	dept_name varchar(30));
+	dept_no varchar(4) primary key not null,
+	dept_name varchar(30) not null);
 
 create table dept_emp (
-	emp_no int,dept_no varchar(4),
-	constraint fk_employee foreign key(emp_no) references employees(emp_no),
-	constraint fk_department foreign key(dept_no) references departments(dept_no));
+	emp_no int not null,
+	dept_no varchar(4) not null,
+	foreign key(emp_no) references employees(emp_no),
+	foreign key(dept_no) references departments(dept_no));
 
 create table dept_manager (
-	dept_no varchar(4),
-	emp_no int,
-	constraint fk_department foreign key(dept_no) references departments(dept_no),
-	constraint fk_employee foreign key(emp_no) references employees(emp_no));
+	dept_no varchar(4) not null,
+	emp_no int not null,
+	foreign key(dept_no) references departments(dept_no),
+	foreign key(emp_no) references employees(emp_no));